replace dialogWindow widgets with fxml files
