+++ src/ca/bcit/jfx/res/stylesheets/application.cssext
@@ -56,7 +56,7 @@
 
 .accordion {
 	-fx-background-color: radial-gradient(center 50% 50%, radius 100%, #004B9E, #004B9E, #003C79);
-	-fx-background-image: url("./images/LogoEON.png");
+	-fx-background-image: url("../images/LogoEON.png");
 	-fx-background-position: bottom center, center center;
 	-fx-background-repeat: no-repeat;
 	-fx-background-size: contain;
@@ -72,7 +72,7 @@
 }
 
 #splash-screen-pane {
-	-fx-background-image: url("./images/bg.png");
+	-fx-background-image: url("../images/bg.png");
 	-fx-background-size: cover;
 }
 
@@ -86,4 +86,9 @@
 	-fx-font-size: 16px; 
 	-fx-fill: #e5e3e3;
 	-fx-font-weight: bold;
+}
+
+#api-key-dialog {
+	-fx-background-image: url("/ca/bcit/jfx/res/images/bg.png");
+	-fx-background-size: cover;
 }
