+++ src/ca/bcit/jfx/res/stylesheets/dialogWindow.cssext
@@ -1,0 +1,9 @@
+.grid-pane {
+	-fx-background-image: url("/ca/bcit/jfx/res/images/bg.png");
+	-fx-background-size: cover;
+}
+
+.title-text {
+	-fx-font-size: 20;
+	-fx-fill: #e5e3e3;
+}
