+++ src/ca/bcit/jfx/res/views/MainWindow.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 <!--suppress ALL -->
-<GridPane gridLinesVisible="true" prefWidth="1300.0" prefHeight="800.0" stylesheets="/ca/bcit/jfx/res/application.cssext" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ca.bcit.jfx.controllers.MainWindowController">
+<GridPane gridLinesVisible="true" prefWidth="1300.0" prefHeight="800.0" stylesheets="/ca/bcit/jfx/res/stylesheets/application.cssext" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ca.bcit.jfx.controllers.MainWindowController">
 	<!-- TO HOLD THE MAP -->
 	<AnchorPane id="AnchorPane" maxHeight="1500" maxWidth="1500" prefWidth="1000" style="-fx-background-color: azure;" >
 		<children>
