+++ src/ca/bcit/jfx/res/views/SplashScreen.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Text?>
 <?import javafx.scene.control.ProgressIndicator?>
 
-<AnchorPane fx:id="splashScreenPane" id="splash-screen-pane" stylesheets="/ca/bcit/jfx/res/application.cssext" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
+<AnchorPane fx:id="splashScreenPane" id="splash-screen-pane" stylesheets="/ca/bcit/jfx/res/stylesheets/application.cssext" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
 	<children>
 		<ImageView fitHeight="100.0" fitWidth="100.0" layoutX="30.0" layoutY="30.0" pickOnBounds="true" preserveRatio="true">
 			<image>
